Move control logic out of model
